# Better startup performance