### What

I've dropped most uses of innerHTML, as textContent is like 5x faster for these tiny text-only strings. The prevArrow/nextArrow are still using innerHTML, due to the inline SVGs. It'd be nice to avoid innerHTML here, but that's a larger refactor.

I've also removed the default parameters for `createElement` as the compiled version relied on `arguments` which isn't very performant in a hot path.

### Results

On my macbook air, I'm seeing the times of `build()` go **from ~9ms to ~2ms**.
For the demo page, the full startup has gone **from ~400ms to ~230ms**.

So, let's average that out and say we're **~300% faster**. 🔥 


### Seeing it

Here's a before and after trace: https://chromedevtools.github.io/timeline-viewer/?loadTimelineFromURL=https://www.dropbox.com/s/7238b2rphoof6j7/flatpickr-before.json?dl=0,%20https://www.dropbox.com/s/byyqvjqzqjj37ur/flatpickr-after.json?dl=0

Screenshot:
![image](https://cloud.githubusercontent.com/assets/39191/21597170/0ed180e0-d114-11e6-9f77-5eaffbbded80.png)

(You'll see the big forced layouts in `positionCalendar`.. It may be possible to avoid those, but luckily, it only hits during startup for the `inline` configuration.)