@chmln 2017-01-02
Hello Paul,

Thanks for the contribution!

Performance is one of the main, continuous goals of this project.
I'm also aware of a few other areas that need optimization, e.g. `positionCalendar()` in `open()` and its dependency on expensive `getBoundingClientRect()`. Those will be addressed in near-future releases.

Best regards.